rag-test-1/requirements.txt
langchain==0.1.17
langchain-google-genai==0.0.8
google-generativeai==0.8.5
faiss-cpu==1.8.0
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1